chalice package
