+++ package/.chalice/config.json
@@ -1,0 +1,11 @@
+{
+  "version": "2.0",
+  "app_name": "package",
+  "manage_iam_role": false,
+  "iam_role_arn": "arn:aws:iam::862902209576:role/lambda_list_accounts",
+  "stages": {
+    "dev": {
+      "api_gateway_stage": "api"
+    }
+  }
+}
