+++ package/.gitignore
@@ -1,0 +1,2 @@
+.chalice/deployments/
+.chalice/venv/
