+++ package/requirements.txt
@@ -1,0 +1,2 @@
+matplotlib==3.0.2
+numpy==1.16.0
